Add cook mode
